requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
numpy==1.21.2
Pillow==8.3.2
scipy==1.7.1
